% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{c_ijkpos}
\alias{c_ijkpos}
\title{Convert physical coordinates to pixel coordinates}
\usage{
c_ijkpos(xyz, origin, voxdims, dims, clamp = FALSE)
}
\arguments{
\item{xyz}{Nx3 matrix of physical coordinates}

\item{origin}{Numeric: 3d coordinates of the origin}

\item{voxdims}{Numeric: 3 numbers describing the voxel dimensions}

\item{dims}{Integer dimensions of the 3d image array}

\item{clamp}{Logical: whether or not to clamp values within the pixel
boundaries of the image.}
}
\value{
Nx3 integer matrix of pixel coordinates
}
\description{
Convert physical coordinates to pixel coordinates
}
