% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{c_listlengths}
\alias{c_listlengths}
\title{A simple function to compute the lengths of the elements of an R list}
\usage{
c_listlengths(L)
}
\arguments{
\item{L}{a list}
}
\value{
An integer vector containing the length of each element of \code{L}
}
\description{
A simple function to compute the lengths of the elements of an R list
}
\details{
This is equivalent to the \code{base::lengths} however it it much
faster for long lists (and somewhat slower for short ones).
}
