% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{c_seglengths}
\alias{c_seglengths}
\alias{c_total_cable}
\title{Compute summed segment lengths or total cable}
\usage{
c_seglengths(sl, x, y, z)

c_total_cable(sl, x, y, z)
}
\arguments{
\item{sl}{A \code{seglist} with 1-indices into vectors x,y,z}

\item{x, y, z}{Numeric vectors with 3D coordinate data
(which could be columns from a data frame)}
}
\description{
\code{c_seglengths} comutes the summed segment length equivalent
 to \code{nat::seglengths(sumsegment = T)}

\code{c_total_cable} computes the summed total cable for a
whole neuron. It's intended use is the \code{nat::summary.neuron} function.
}
