% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{c_sub2ind}
\alias{c_sub2ind}
\title{Find 1D index given n-dimensional indices}
\usage{
c_sub2ind(dims, indices)
}
\arguments{
\item{dims}{Integer dimensions of the array (usually 3d)}

\item{indices}{Nx3 integer matrix of pixel coordinates}
}
\value{
numeric vector of linear indices into the array
}
\description{
Find 1D index given n-dimensional indices
}
