% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{c_topntail}
\alias{c_topntail}
\title{Find the first and last element of a list}
\usage{
c_topntail(L)
}
\arguments{
\item{L}{a list containing integer vectors}
}
\value{
An integer vector containing the length of each element of \code{L}
}
\description{
Find the first and last element of a list
}
