% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{c_topntail}
\alias{c_topntail}
\alias{c_topntail_list}
\title{Find the first and last elements of all vectors in a list}
\usage{
c_topntail(L)

c_topntail_list(L)
}
\arguments{
\item{L}{a list containing integer vectors, typically a \code{seglist}}
}
\value{
For \code{c_topntail} an integer \code{matrix}. For \code{c_topntail_list}
  a \code{list}.
}
\description{
\code{c_topntail} returns an Nx2 matrix containing the start and end
of each of the vectors in the input list. Length 0 vectors are ignored, while
length 1 vectors are duplicated

For \code{c_topntail_list}, a list of the same length as
\code{L} having the same elements when their length is <=2 or
the first and last elements when length>2.
}
