% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/natcpp-package.R
\docType{package}
\name{natcpp-package}
\alias{natcpp}
\alias{natcpp-package}
\title{natcpp: Fast C++ Primitives for the 'NeuroAnatomy Toolbox'}
\description{
Fast functions implemented in C++ via 'Rcpp' to support the
    'NeuroAnatomy Toolbox' ('nat') ecosystem. These functions provide large
    speed-ups for basic manipulation of neuronal skeletons over pure R
    functions found in the 'nat' package. The expectation is that end
    users will not use this package directly, but instead the 'nat'
    package will automatically use routines from this package when it is
    available to enable large performance gains.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/natverse/natcpp}
  \item Report bugs at \url{https://github.com/natverse/natcpp/issues}
}

}
\author{
\strong{Maintainer}: Gregory Jefferis \email{jefferis@gmail.com} (\href{https://orcid.org/0000-0002-0587-9355}{ORCID})

}
\keyword{internal}
